Refactor: move error formatting to shared utils as error_utils
Follow-up review feedback: the helper is generic enough to be reused by other
error handling, so it belongs in the shared utils directory under a name that
does not read as MCP-only.

- core/src/tools/mcp/mcp_error_utils.ts -> core/src/utils/error_utils.ts,
  alongside case_utils/file_utils/failover_utils; test moves to
  core/test/utils/error_utils_test.ts to match the sibling convention.
- Drop the MCP framing: the module doc now describes generic error formatting,
  and StreamableHTTPError is mentioned only as one of several supported error
  shapes (`.status`, `.response`, numeric `.code`) rather than the purpose.
- logTransportError stays in the MCP layer (now a private function in
  mcp_session_manager.ts) so the 'MCP transport error: ' label is not baked
  into a generic module; error_utils.ts exports only formatError.

Still internal: nothing added to index.ts/common.ts. Pure move/rename with no
behavior change, no any/eslint-disable. 52 targeted tests pass with 100%
line+branch coverage of error_utils.ts and mcp_session_manager.ts.
